pyproject.toml
[tool.poetry]
name = "llm_cli"
version = "0.6.2"
description = "CLI tool for LLM interactions"
authors = ["Andres Crucetta"]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.7"
rich = "^13.7.0"
requests = "^2.31.0"
pyyaml = "^6.0.1"
google-genai = "^1.7.0"
openai = "^1.66.5"
prompt_toolkit = "^3.0.41"

[tool.poetry.scripts]
llm = "llm_cli.main:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
